+++ l10n_th_gov_work_acceptance/__manifest__.py
@@ -10,14 +10,13 @@
     "category": "Localization / Purchase",
     "depends": [
         "hr",
-        "purchase_work_acceptance",
+        "base_tier_validation_formula",
+        "base_tier_validation_forward",
+        "base_tier_validation_server_action",
         "purchase_work_acceptance_evaluation",
         "purchase_work_acceptance_late_fines",
-        "hr_expense_work_acceptance",
-        "hr_expense_from_purchase_request",
-        "hr_expense_advance_clearing",
-        "l10n_th_gov_purchase_request",
         "purchase_work_acceptance_tier_validation",
+        "l10n_th_gov_purchase_request",
     ],
     "data": [
         "security/ir.model.access.csv",
@@ -26,7 +25,6 @@
         "data/tier.definition.csv",
         "views/work_acceptance_views.xml",
         "views/purchase_views.xml",
-        "views/expense_views.xml",
     ],
     "installable": True,
     "development_status": "Beta",
