+++ l10n_th_gov_work_acceptance_expense/models/__init__.py
@@ -1,4 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
-from . import work_acceptance_committee
+from . import expense
 from . import work_acceptance
-from . import purchase
