+++ l10n_th_gov_work_acceptance_expense/readme/DESCRIPTION.rst
@@ -1,0 +1,7 @@
+This module contain process related to purchase and expense's work acceptance, including,
+
+* Only reduced step purchase request is eligible in expense sheet.
+* WA acceptance is required if purchase request is selected.
+* Add Committee tab to WA document, and if any, pull committee from purchase request or expense created from purchase request
+* Work Acceptance Evaluation with criterias
+* Work Acceptance Late Fines, on the last acceptance.
