+++ l10n_th_gov_work_acceptance_expense/security/ir.model.access.csv
@@ -1,0 +1,4 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_work_acceptance_committee,procurement.committee,l10n_th_gov_work_acceptance.model_work_acceptance_committee,hr_expense_work_acceptance.group_enable_wa_on_exp,1,1,1,1
+access_work_acceptance_evaluation_user,access_work_acceptance_evaluation_manager,purchase_work_acceptance_evaluation.model_work_acceptance_evaluation,hr_expense_work_acceptance.group_enable_wa_on_exp,1,0,0,0
+access_work_acceptance_evaluation_score_user,access_work_acceptance_evaluation_score_manager,purchase_work_acceptance_evaluation.model_work_acceptance_evaluation_score,hr_expense_work_acceptance.group_enable_wa_on_exp,1,0,0,0
